refactor: EventPublisher 공통 로직으로 분리
