+++ src/main/java/com/flytrap/rssreader/api/post/business/event/PostOpenEvent.java
@@ -1,4 +1,4 @@
-package com.flytrap.rssreader.api.post.business.event.postOpen;
+package com.flytrap.rssreader.api.post.business.event;
 
 import com.flytrap.rssreader.api.member.domain.AccountId;
 import com.flytrap.rssreader.api.post.domain.PostAggregate;
